requirements.txt
# ── API Framework ──
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# ── PDF Processing ──
pdfplumber==0.11.4
pdf2image==1.17.0
pytesseract==0.3.13

# ── AI/ML ──
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1

# ── PDF Report Generation ──
fpdf2==2.8.2

# ── Config & Utilities ──
pydantic-settings==2.7.1
python-dotenv==1.0.1

# ── Testing ──
pytest==8.3.4
httpx==0.28.1
pytest-asyncio==0.24.0
